fix: support non-integer zoom levels (#421)
fix(PTZPositionDegree): Support non-integer zoom levels

The zoom level reported by the protect API isn't always an integer.
Update the type to float to support this case.

Fixes #420

+++ src/uiprotect/data/types.py
@@ -679,7 +679,7 @@ RepeatTimes = Annotated[int, Field(ge=1, le=6)]
 class PTZPositionDegree(BaseModel):
     pan: float
     tilt: float
-    zoom: int
+    zoom: float
 
 
 class PTZPositionSteps(BaseModel):
